Add files via upload
experimental data in .h5 format can be opened via hdf5.
Keys: 'time_ns', 'population', 'population_off_resonance'
Please refer to the png file for the visualization of the data. The off resonant case is 100 MHz detuned.
Ipython file has all the simulation codes used, the link to the data is not included. ZK